+++ breakout/engine/src/game.cpp
@@ -1,7 +1,7 @@
 #include "engine/game.hpp"
 #include <iostream>
 
-Game::Game(const unsigned int width, const unsigned int height, const std::string& _name) : State(GAME_MENU), Keys(), Width(width), Height(height), name(_name) {}
+Game::Game(const int width, const int height, const std::string& _name) : State(GAME_MENU), Keys(), Width(width), Height(height), name(_name) {}
 
 Game::~Game() {}
 
@@ -11,9 +11,9 @@ void Game::SetEngineDelegate(Engine* engine) {
 
 void Game::Init() {}
 
-void Game::Update(float dt) {}
+void Game::Update(double dt) {}
 
-void Game::ProcessInput(float dt) {}
+void Game::ProcessInput(double dt) {}
 
 void Game::pressed(int key) {}
